Внес правки по замечаниям после ревью
