+++ Tracker/CoreData/TrackerCategoryStore.swift
@@ -22,9 +22,6 @@ final class TrackerCategoryStore: NSObject {
     private let context: NSManagedObjectContext
     private let trackerStore = TrackerStore()
     private var insertedIndexes: IndexSet?
-    private var deletedIndexes: IndexSet?
-    private var updatedIndexes: IndexSet?
-    private var movedIndexes: Set<StoreUpdate.Move>?
     
     private lazy var fetchedResultsController: NSFetchedResultsController<TrackerCategoryCoreData> = {
         let fetchRequest = NSFetchRequest<TrackerCategoryCoreData>(entityName: "TrackerCategoryCoreData")
@@ -45,8 +42,8 @@ final class TrackerCategoryStore: NSObject {
     
     // MARK: - Inits
     convenience override init() {
-        let appDelegate = UIApplication.shared.delegate as! AppDelegate
-        let context = appDelegate.persistentContainer.viewContext
+        let container = PersistentContainer.shared
+        let context = container.persistentContainer.viewContext
         self.init(context: context)
     }
     
@@ -94,13 +91,12 @@ final class TrackerCategoryStore: NSObject {
     
     // MARK: - Private Methods
     private func saveContext() {
-        if context.hasChanges {
-            do {
-                try context.save()
-            } catch {
-                let nserror = error as NSError
-                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
-            }
+        guard context.hasChanges else { return }
+        do {
+            try context.save()
+        } catch {
+            let nserror = error as NSError
+            fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
         }
     }
     
@@ -122,26 +118,9 @@ extension TrackerCategoryStore: NSFetchedResultsControllerDelegate {
                     newIndexPath: IndexPath?
     ) {
         switch type {
-        case .delete:
-            if let indexPath = indexPath {
-                deletedIndexes?.insert(indexPath.item)
-            }
         case .insert:
             if let indexPath = newIndexPath {
                 insertedIndexes?.insert(indexPath.item)
-            }
-        case .move:
-            if let oldIndexPath = indexPath,
-               let newIndexPath = newIndexPath {
-                movedIndexes?.insert(
-                    .init(oldIndex: oldIndexPath.item,
-                          newIndex: newIndexPath.item
-                         )
-                )
-            }
-        case .update:
-            if let indexPath = indexPath {
-                updatedIndexes?.insert(indexPath.item)
             }
         default:
             break
