+++ Tracker/Presentation/ScheduleViewController.swift
@@ -108,7 +108,7 @@ final class ScheduleViewController: UIViewController {
 // MARK: - TableViewDataSource
 extension ScheduleViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return weekDays.count
+        weekDays.count
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
@@ -149,6 +149,6 @@ extension ScheduleViewController: UITableViewDataSource {
 // MARK: - TableViewDelegate
 extension ScheduleViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 75
+        75
     }
 }
